Fixin structure for services input, update instance for customer-service on open-feign

+++ shopping/src/main/java/acc/microservices/shopping/repository/InvoiceRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import acc.microservices.shopping.model.Invoice;
 
 @Repository
-public interface InvoiceRepository extends JpaRepository<Invoice, Long> {
+public interface InvoiceRepository extends JpaRepository<Invoice, Integer> {
   public List<Invoice> findByCustomerId(Long customerId);
 
   public Optional<Invoice> findByNumberInvoice(String numberInvoice);
